Added Shortcuts, Improved UI

+++ src/hooks/Utils/useKeybinds.ts
@@ -1,31 +1,29 @@
+import useDeletePoint from "../Point/useDeletePoint.ts";
+import { useSelectedPointValue } from "../Point/useSelectPoint.ts";
+import { useRedoPath, useUndoPath } from "./useUndoHistory.ts";
+import React from "react";
+
 export default function useKeybinds() {
-    /*
     const deletePoint = useDeletePoint();
     const selectedPointID = useSelectedPointValue();
+    const [, undo] = useUndoPath();
+    const [, redo] = useRedoPath();
+    
+    const onKeyDown = React.useCallback((e: KeyboardEvent) => {
+        if (e.key === "Delete" && selectedPointID)
+            deletePoint(selectedPointID);
+        if (e.key === "z" && e.ctrlKey)
+            undo();
+        if (e.key === "y" && e.ctrlKey)
+            redo();
+    }, [deletePoint, selectedPointID, undo, redo]);
 
-    const hotkeys = React.useMemo<HotkeyConfig[]>(() => [
-        {
-            combo: "backspace",
-            global: true,
-            label: "Delete Point",
-            onKeyDown: () => {
-                if (selectedPointID)
-                    deletePoint(selectedPointID);
-            }
-        },
-        {
-            combo: "delete",
-            global: true,
-            label: "Delete Point",
-            onKeyDown: () => {
-                if (selectedPointID)
-                    deletePoint(selectedPointID);
-            }
+    React.useEffect(() => {
+        window.addEventListener("keydown", onKeyDown);
+        return () => {
+            window.removeEventListener("keydown", onKeyDown);
         }
-    ], [deletePoint, selectedPointID]);
+    }, [onKeyDown]);
 
-    return useHotkeys(hotkeys);
-    */
-    // TODO: Implement useHotkeys
     return null;
 }
